Clarified role of "mimeType"
There's a balance to be struck here. It is useful to be able to provide
MIME type hints, so user agents can present an interface that allows
the user to anticipate what is at the other end of the link.  On the
other hand, we don't want to allow these type hints to re-interpret
safe data.  I might want to include a plain-text illustration of some
malicious code, without worrying that somebody could link to
"/tutorial/security#/example/code", re-interpreting it as a client-
side script that would then execute in the context of my domain.